requirements.txt
sox
absl-py
yapf
pylint
flake8
tqdm
sentencepiece
librosa
jieba
pandas<=0.25.1 
kaldiio
tarfile
rarfile
inflect
urllib
tempfile
https://github.com/kpu/kenlm/archive/master.zip
scipy
sklearn